Try reading ice saturation from STK
Couldn't make the ACE Ice Saturation field work correctly
the way it was implemented, no matter what I tried.

Instead, try to read it the same way that the boundary
indicator is read. There are issues with the exodus files
for the thermal and mechanical problems. They appear
to be out of sync in the initial implementation. But
when syncing them correctly, the initial mechanical
exodus file, which is the last thermal exodus output,
is incomplete, despite the thermal problem having run.
This seems to be an issue with a file buffer, but can't
figure out yet how to force a flush of that buffer,
so that the thermal output file is fully written
for the mechanical problem to read.